Added the ability to cancel orders, clients can now cancel orders

+++ src/main/java/com/glfx/orderprocessingservice/service/OrderService.java
@@ -2,11 +2,13 @@ package com.glfx.orderprocessingservice.service;
 
 import com.glfx.orderprocessingservice.exceptions.InvalidOrderException;
 import com.glfx.orderprocessingservice.exceptions.OrderNotFoundException;
-import com.glfx.orderprocessingservice.market.marketDTO.MarketData;
 import com.glfx.orderprocessingservice.model.Order;
+import com.glfx.orderprocessingservice.model.OrderToExchange;
 import com.glfx.orderprocessingservice.repository.OrderRepository;
 import com.glfx.orderprocessingservice.utils.OrderType;
 import com.glfx.orderprocessingservice.utils.Side;
+import com.glfx.orderprocessingservice.utils.Status;
+import com.glfx.orderprocessingservice.validators.OrderValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -29,22 +31,134 @@ public class OrderService {
     private OrderRepository orderRepository;
 
     @Autowired
-    MarketData marketData;
+    private OrderValidator orderValidator;
 
-    public List<Order> getAllPortfolioOrders(Long portfolioID){
-        //List<Order> orders =  orderRepository.findAll();
-        return orderRepository.findByPortfolioID(portfolioID);
+    @Autowired
+    private OrderToExchange orderToExchange;
+
+
+    public List<Order> getAllOrders(){
+        return orderRepository.findAll();
+    }
+
+
+    public void createOrder(Order order) throws InvalidOrderException {
+
+        if(orderValidator.validate(order)){
+            if(order.getType().equalsIgnoreCase(OrderType.LIMIT.toString())){
+                orderToExchange.setProduct(order.getProduct());
+                orderToExchange.setSide(order.getSide());
+                orderToExchange.setPrice(order.getPrice());
+                orderToExchange.setType(order.getType());
+                orderToExchange.setQuantity(order.getQuantity());
+
+                if(orderValidator.getExchanges().equalsIgnoreCase("exchange1")){
+
+                    String response = webclient.build().post()
+                            .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
+                            .body(Mono.just(orderToExchange), OrderToExchange.class)
+                            .retrieve()
+                            .bodyToMono(String.class)
+                            .block();
+                    order.setOrderIdFromExchange(response.substring(1,response.length()-1));
+                    order.setStatus(Status.NOT_EXECUTED.toString());
+                    orderRepository.save(order);
+                }
+
+                else if(orderValidator.getExchanges().equalsIgnoreCase("exchange2")){
+
+                    String response = webclient.build().post()
+                            .uri(exchange2+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
+                            .body(Mono.just(orderToExchange), OrderToExchange.class)
+                            .retrieve()
+                            .bodyToMono(String.class)
+                            .block();
+                    order.setStatus(Status.NOT_EXECUTED.toString());
+                    order.setOrderIdFromExchange(response.substring(1,response.length()-1));
+                    orderRepository.save(order);
+                }
+
+                else {
+
+                    System.out.println("Splitting order will be done here!");
+                    String response = webclient.build().post()
+                            .uri(exchange2+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
+                            .body(Mono.just(orderToExchange), OrderToExchange.class)
+                            .retrieve()
+                            .bodyToMono(String.class)
+                            .block();
+                    order.setStatus(Status.NOT_EXECUTED.toString());
+                    order.setOrderIdFromExchange(response.substring(1,response.length()-1));
+                    orderRepository.save(order);
+                }
+            }
+
+            else {
+                    orderToExchange.setProduct(order.getProduct());
+                    orderToExchange.setSide(order.getSide());
+                    orderToExchange.setType(order.getType());
+                    orderToExchange.setQuantity(order.getQuantity());
+
+                    if(orderValidator.getExchanges().equalsIgnoreCase("exchange1")){
+
+                        String response = webclient.build().post()
+                                .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
+                                .body(Mono.just(orderToExchange), OrderToExchange.class)
+                                .retrieve()
+                                .bodyToMono(String.class)
+                                .block();
+                        order.setStatus(Status.NOT_EXECUTED.toString());
+                        order.setOrderIdFromExchange(response.substring(1,response.length()-1));
+                        orderRepository.save(order);
+                    }
+
+                    else if(orderValidator.getExchanges().equalsIgnoreCase("exchange2")){
+
+                        String response = webclient.build().post()
+                                .uri(exchange2+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
+                                .body(Mono.just(orderToExchange), OrderToExchange.class)
+                                .retrieve()
+                                .bodyToMono(String.class)
+                                .block();
+                        order.setStatus(Status.NOT_EXECUTED.toString());
+                        order.setOrderIdFromExchange(response.substring(1,response.length()-1));
+                        orderRepository.save(order);
+                    }
+
+                    else {
+
+                        System.out.println("Splitting order will be done here!");
+                        String response = webclient.build().post()
+                                .uri(exchange2+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
+                                .body(Mono.just(orderToExchange), OrderToExchange.class)
+                                .retrieve()
+                                .bodyToMono(String.class)
+                                .block();
+                        order.setStatus(Status.NOT_EXECUTED.toString());
+                        order.setOrderIdFromExchange(response.substring(1,response.length()-1));
+                        orderRepository.save(order);
+                    }
+            }
+
+
+        }
+
+        else{
+            throw new InvalidOrderException("Invalid Order");
+        }
+
     }
 
     public Optional<Order> checkOrderStatus(Long id) throws OrderNotFoundException{
-        Order response = webclient.build()
-                .get()
-                .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order/"+id)
-                .retrieve()
-                .bodyToMono(Order.class)
-                .block();
+
         Optional<Order> order = orderRepository.findById(id);
         if (order.isPresent()) {
+            OrderToExchange response = webclient.build()
+                .get()
+                .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order/"+order.get().getOrderIdFromExchange())
+                .retrieve()
+                .bodyToMono(OrderToExchange.class)
+                .block();
             return order;
         }
         else{
@@ -52,77 +166,16 @@ public class OrderService {
         }
     }
 
-    public void marketOrder(Order order){
-        if(order.getSide() == Side.Buy){
-            orderRepository.save(order);
-        }
-        else if (order.getSide() == Side.Sell) {
-            orderRepository.save(order);
-        }
-    }
-
-    public void limitOrder(Order order) throws InvalidOrderException{
-        if(order.getQuotedPrice() <= 0) {
-            throw new InvalidOrderException("Quoted price must be greater than 0 for limit orders!");
-        }
-        else {
-            if (order.getSide() == Side.Buy) {
-                orderRepository.save(order);
-            }
-            else if (order.getSide() == Side.Sell) {
-                orderRepository.save(order);
-            }
-        }
-    }
-
-    public void sellOrder(Order order) throws InvalidOrderException {
-        if(order.getType() == OrderType.Limit_Order){
-            limitOrder(order);
-        }
-        else if (order.getType() == OrderType.Market_Order) {
-            marketOrder(order);
-        }
-    }
-
-    public void buyOrder(Order order) throws InvalidOrderException {
-        if(order.getType() == OrderType.Limit_Order){
-            limitOrder(order);
-        }
-        else if (order.getType() == OrderType.Market_Order){
-            marketOrder(order);
-        }
-    }
-
-    public void createOrder(Order order) throws InvalidOrderException {
-        if(order.getQuantity() <= 0){
-            throw new InvalidOrderException("Stock quantity must be greater than zero to place an order!");
-        }
-        else {
-            if (order.getSide() == Side.Buy) {
-                buyOrder(order);
-            } else if (order.getSide() == Side.Sell) {
-                sellOrder(order);
-            }
-            String response = webclient.build().post()
-                    .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order")
-                    .body(Mono.just(order), Order.class)
-                    .retrieve()
-                    .bodyToMono(String.class)
-                    .block();
-            //order.setId(response.substring(1,response.length()-1));
-            orderRepository.save(order);
-        }
-    }
 
     public void updateOrder(Long id, Order order) throws OrderNotFoundException {
         Optional<Order> orderDB = orderRepository.findById(id);
-        boolean response = webclient.build()
-                .put()
-                .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order/"+id)
-                .body(Mono.just(order), Order.class)
-                .retrieve()
-                .bodyToMono(Boolean.class)
-                .block();
+//        boolean response = webclient.build()
+//                .put()
+//                .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order/"+id)
+//                .body(Mono.just(order), Order.class)
+//                .retrieve()
+//                .bodyToMono(Boolean.class)
+//                .block();
         if (orderDB.isPresent()) {
             orderDB = Optional.of(order);
             orderRepository.save(orderDB.get());
@@ -133,13 +186,20 @@ public class OrderService {
 
     }
 
-    public void cancelOrder(Long id){
-        boolean response = webclient.build()
-                .delete()
-                .uri(exchange1+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order/"+id)
-                .retrieve()
-                .bodyToMono(Boolean.class)
+    public void cancelOrder(Long id) throws OrderNotFoundException {
+        Optional<Order> order = orderRepository.findById(id);
+
+        if(order.isPresent()){
+                boolean response = webclient.build()
+                        .delete()
+                        .uri(exchange2+"/15a204cd-9f59-45e9-8908-fbfe7f20480d/order/"+order.get().getOrderIdFromExchange())
+                        .retrieve()
+                        .bodyToMono(Boolean.class)
                         .block();
-        orderRepository.deleteById(id);
+            }
+        else throw new OrderNotFoundException("Can't find order with ID " + id);
+
+        order.get().setDeleted(true);
+        orderRepository.delete(order.get());
     }
 }
